refactor(theme): move types to common type file
