+++ src/framework/theme/component/theme/index.ts
@@ -1,4 +1,3 @@
-export * from './type';
 export * from './themeContext';
 export {
   ThemeProvider,
