+++ src/framework/theme/component/theme/theme.spec.config.ts
@@ -1,4 +1,6 @@
-export const theme = {
+import { ThemeType } from '../../type';
+
+export const theme: ThemeType = {
   grayLight: '#E0E0E0',
   grayPrimary: '#9E9E9E',
   grayDark: '#616161',
@@ -10,7 +12,7 @@ export const theme = {
   tealDark: '#00796B',
 };
 
-export const themeInverse = {
+export const themeInverse: ThemeType = {
   grayLight: '#616161',
   grayPrimary: '#9E9E9E',
   grayDark: '#E0E0E0',
