--- src/framework/theme/component/theme/type.ts
@@ -1,6 +1,0 @@
-// TODO(theme/type): declare Theme types
-
-export type ThemeType = any;
-export type StyleType = any;
-export type ThemedStyleType = any;
-export type StyleSheetType = any;
